chore: export types from logger module (#7292)
