+++ .changeset/eighty-turkeys-marry.md
@@ -1,0 +1,5 @@
+---
+"@vue-storefront/sdk": patch
+---
+
+Export types from sdk logger module
